Added "Install" and "Remove" screens to SimpleBasisAMS.  Users can now install content without having to run the "runinstall" script first. The UI screens guide users on how to install.  Read the runjump script for details on how to set a value for the provisioning server.
Added a blank preferences screen with the intention of adding something 
in the future.

Created the new InstallerTool presentation mode.  This tool is run with 
the new runxinstall script.  The InstallerTool presentation mode and its 
associated runxinstall script is different than the runinstall script 
because the InstallerTool runs on behalf of the JUMP Executive, whereas 
runinstall does not.  The difference can affect modules used in the 
install process that require the JUMP Executive to be running.  For 
example, PUSH needs this support.  Please read runxinstall for details 
on how to run.  Unfortunately, when running within the JUMP Executive, 
you currently cannot interact with the shell, so InstallerTool does not 
support user input in the same fashion as in runinstall.

Temporarily commenting out the question given to the user when a midlet 
or xlet is part of a suite with multiple applications.  The current 
default will be to remove all applications in the suite without asking 
the user.  This is to support the InstallerTool presentation mode, which 
runs in the JUMP executive and is currently unable to provide an 
interface for this question.  This topic will be revisited shortly.