Until the full rotation support is integrated there is a need to implement some solution to allow display of ams system displays in orientation consistent with the current running application (midlet).
Displays owned by MIDlet instances are initialized to be shown in portrait orientation. Displays owned by objects that are not MIDlet instances (system displays) are initialized to be shown in the current device orientation mode. When application Display instance gains foreground it may change device orientation depending on its orientation mode. I.e. if the current device orientation is landscape and Display was initialized to be shown in portrait mode then the device orientation changes and set to portrait mode. System displays never change device orientation.

Reviewed by Steve.